MCMC fitting routine
Uses markov chain monte carlo fitting technique to fit either 1 or 2 components to the [OIII] doublet 
in each spatial pixel. First it tries to fit 2 components, then evaluates the S/N ratio to determine if 
there is statistical support for the existance of a second component. If not, then refit but with only 
one component. Appends the best fit parameters and uncertainties to the datacube file.